refactor: Extract load_recipes function to utils module for better organization
